comment code + todo updated + unsigned to unsigned short

+++ include/damerau_levenshtein.hh
@@ -2,6 +2,6 @@
 
 using namespace std;
 
-vector<tuple<string, unsigned, unsigned>> distance_dl(Trie*& t, const string& word, unsigned max_dist);
-vector<tuple<string, unsigned, unsigned>> distance_rec(Trie* t, char c, const string& word,
-		unsigned max_dist, vector<unsigned> previous_row, string current_word);
+vector<tuple<string, unsigned short, unsigned>> distance_dl(Trie*& t, const string& word, unsigned short max_dist);
+vector<tuple<string, unsigned short, unsigned>> distance_rec(Trie* t, char c, const string& word,
+		unsigned short max_dist, vector<unsigned short> previous_row, string current_word);
